We don't need to run a double check on the token signature because auth-verify derives the IKM from it, and if the data is pulled from the cache on the server and auth::verify passes and the HMAC is the same then it effectively validates that the token that signed the request, is the same one the server is aware of.
This commit removes the extra check for the constant time comparison
between the signature public key provided by the client, and the
signature public key stored with the server.

The threat model isn't meant to be resistent to an adversary who has
direct access to the server, as that individual can extract and read and
decrypt all messages as-is - there is no extra security benefit to
running a double check then previous checks are sufficient.